refactor: consolidate task components and remove unused styles
