+++ app/routes/01-lets-start/route.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { Box } from '~/components/common/box/box';
 import { ConfettiFx } from '~/components/fx/confetti-fx/confetti-fx';
 import { CursorFx } from '~/components/fx/cursor-fx/cursor-fx';
-import { Task01 as Task } from '~/components/tasks/01/task-01';
+import { Task01 as Task } from '~/components/tasks/task-01';
 import styles from './lesson-01.module.scss';
 import { LessonsFooter } from '~/components/lessons-footer/lessons-footer';
 import { LESSON_02 } from '~/router/config';
