+++ app/routes/02-elements-panel/route.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Box } from '~/components/common/box/box';
 import { ConfettiFx } from '~/components/fx/confetti-fx/confetti-fx';
-import { Task02 as Task } from '~/components/tasks/02/task-02';
+import { Task02 as Task } from '~/components/tasks/task-02';
 import styles from './lesson-02.module.scss';
 import { LessonsFooter } from '~/components/lessons-footer/lessons-footer';
 import { LESSON_01, LESSON_03 } from '~/router/config';
