+++ app/routes/06-create-class/route.tsx
@@ -3,7 +3,7 @@ import { useCallback, useState } from 'react';
 import vars from '~/globals/variables.module.scss';
 import { ConfettiFx } from '~/components/fx/confetti-fx/confetti-fx';
 import { useRequestAnimationFrame } from '~/components/hooks';
-import { Task06 as Task } from '~/components/tasks/06/task-06';
+import { Task06 as Task } from '~/components/tasks/task-06';
 import styles from './lesson-06.module.scss';
 import { LessonsFooter } from '~/components/lessons-footer/lessons-footer';
 import { LESSON_05, LESSON_07 } from '~/router/config';
