+++ app/routes/09-grid-layouts/route.tsx
@@ -3,7 +3,7 @@ import type { ColorName } from '~/globals/colors';
 import { Box } from '~/components/common/box/box';
 import { ConfettiFx } from '~/components/fx/confetti-fx/confetti-fx';
 import { useRequestAnimationFrame } from '~/components/hooks';
-import { Task09 as Task } from '~/components/tasks/09/task-09';
+import { Task09 as Task } from '~/components/tasks/task-09';
 import styles from './lesson-09.module.scss';
 import { LessonsFooter } from '~/components/lessons-footer/lessons-footer';
 import { LESSON_08, LESSON_10 } from '~/router/config';
