+++ app/routes/10-work-with-git/route.tsx
@@ -1,5 +1,5 @@
 import Cloud from '../../../src/assets/cloud.svg?react';
-import { Task10 as Task } from '~/components/tasks/10/task-10';
+import { Task10 as Task } from '~/components/tasks/task-10';
 import styles from './lesson-10.module.scss';
 import { LessonsFooter } from '~/components/lessons-footer/lessons-footer';
 import { LESSON_09 } from '~/router/config';
