--- src/components/tasks/01/task-01.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
