--- src/components/tasks/02/task-02.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
