--- src/components/tasks/03/task-03.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
