--- src/components/tasks/04/task-04.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
