--- src/components/tasks/05/task-05.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
