--- src/components/tasks/06/task-06.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
