--- src/components/tasks/07/task-07.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
