--- src/components/tasks/08/task-08.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
