--- src/components/tasks/09/task-09.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
