--- src/components/tasks/10/task-10.module.scss
@@ -1,1 +1,0 @@
-@use '../task.module.scss';
