+++ src/components/tasks/task-05.tsx
@@ -1,6 +1,6 @@
-import styles from './task-05.module.scss';
-import { TaskSymbol } from '../../common/task-symbol/task-symbol';
-import { TaskTag } from '../../common/task-tag/task-tag';
+import styles from './task.module.scss';
+import { TaskSymbol } from '../common/task-symbol/task-symbol';
+import { TaskTag } from '../common/task-tag/task-tag';
 
 export const Task05 = () => {
     return (
